Fix windows burn, take 3
git-svn-id: http://www.virtualbox.org/svn/vbox/trunk@31106 cfe28804-0f27-0410-a406-dd0f0b0b656f

+++ src/VBox/Runtime/r3/socket.cpp
@@ -813,7 +813,7 @@ RTDECL(int) RTSocketWriteNB(RTSOCKET hSocket, const void *pvBuffer, size_t cbBuf
     {
         rtSocketErrorReset();
 #ifdef RT_OS_WINDOWS
-        int    cbNow = RT_MIN(cbToWrite, INT_MAX/2);
+        int    cbNow = RT_MIN((int)cbToWrite, INT_MAX/2);
 #else
         size_t cbNow = cbToWrite;
 #endif
